Remove cancelled channels from pending packets
We can use the link resolution channel in a FIFO to limit the number of
maximum pending resolutions we should queue packets for.

This change also defers starting the goroutine that handles link
resolution completion to when link resolution succeeds, fails or
gets cancelled due to the max number of pending resolutions being
reached.

PiperOrigin-RevId: 351942810